cache package refactored CacheException added ConfigCache converted to interface
