+++ src/main/java/com/configcat/cache/CacheException.java
@@ -1,0 +1,4 @@
+package com.configcat.cache;
+
+public class CacheException extends Exception{
+}
