+++ src/test/java/com/configcat/AutoPollingPolicyTest.java
@@ -1,5 +1,9 @@
 package com.configcat;
 
+import com.configcat.cache.CacheException;
+import com.configcat.cache.ConfigCache;
+import com.configcat.cache.NullConfigCache;
+import com.configcat.cache.SingleValueCache;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
@@ -165,8 +169,8 @@ class AutoPollingPolicyTest {
 
         ConfigCache cache = mock(ConfigCache.class);
 
-        doThrow(new Exception()).when(cache).read(anyString());
-        doThrow(new Exception()).when(cache).write(anyString(), anyString());
+        doThrow(new CacheException()).when(cache).read(anyString());
+        doThrow(new CacheException()).when(cache).write(anyString(), anyString());
 
         PollingMode pollingMode = PollingModes.autoPoll(2);
         ConfigFetcher fetcher = new ConfigFetcher(new OkHttpClient.Builder().build(),
