+++ src/test/java/com/configcat/ConfigFetcherTest.java
@@ -1,5 +1,7 @@
 package com.configcat;
 
+import com.configcat.cache.CacheException;
+import com.configcat.cache.ConfigCache;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
@@ -133,8 +135,8 @@ class ConfigFetcherTest {
         this.server.enqueue(new MockResponse().setResponseCode(200).setBody(TEST_JSON));
         ConfigCache cache = mock(ConfigCache.class);
 
-        doThrow(new Exception()).when(cache).read(anyString());
-        doThrow(new Exception()).when(cache).write(anyString(), anyString());
+        doThrow(new CacheException()).when(cache).read(anyString());
+        doThrow(new CacheException()).when(cache).write(anyString(), anyString());
 
         ConfigFetcher fetcher = new ConfigFetcher(new OkHttpClient.Builder().build(), logger,
                 "", this.server.url("/").toString(), false, PollingModes.manualPoll().getPollingIdentifier());
