+++ src/test/java/com/configcat/LazyLoadingPolicyTest.java
@@ -1,5 +1,8 @@
 package com.configcat;
 
+import com.configcat.cache.ConfigCache;
+import com.configcat.cache.NullConfigCache;
+import com.configcat.cache.SingleValueCache;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import okhttp3.OkHttpClient;
