+++ src/test/java/com/configcat/ManualPollingPolicyTest.java
@@ -1,5 +1,7 @@
 package com.configcat;
 
+import com.configcat.cache.InMemoryCache;
+import com.configcat.cache.NullConfigCache;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import okhttp3.OkHttpClient;
